cleaned up code - now there arent any warnings anymore :)
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@71 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ e16/e/src/snaps.c
@@ -332,7 +332,7 @@ SnapshotEwinDialog(EWin * ewin)
 
 	     s[0] = 0;
 	     slice = 80;
-	     while (i <= strlen(ewin->client.command))
+	     while (i <= (int)strlen(ewin->client.command))
 	       {
 		  last = i;
 		  i += 80;
